fetch2/git.py: drop redundant calls to _contains_ref() from download()
Updating a git repo currently results in multiple calls to
self._contains_ref(), some of which appear to be redundant and
can be eliminated by minor tweaks to the logic in download().

Also drop redundant calls to os.path.exists(ud.clonedir) before
self.need_update(), since need_update() includes it's own built-in
check for the existance of ud.clonedir.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>